+++ CHANGES.rst
@@ -18,6 +18,8 @@
   ``SIA2_PARAMETERS_DESC``. The old names now issue an
   ``AstropyDeprecationWarning``. [#419]
 
+- Registry search now finds SIA v2 services. [#422]
+
 
 1.4 (2022-09-26)
 ================